1.0.224 - Big improvement to search
Discovered problems with album release list processing. Disambiguation was not included in the release title, which would result in a lot less matches because disambiguation is needed to more closely match the service release names.

Max/Min track count had an issue, where it was possible to not have the correct min/max track count, which could cause match failures... This is now resolved.