Add videoDownloadTag to limit artist downloads
Per discussion on #141, this PR adds the ability to specify a videoDownloadTag.
If a value is specified for videoDownloadTag, only artists tagged in Lidarr with
the specified tag will have videos downloaded for them. If videoDownloadTag is
left empty (the default), things work exactly as they did prior to introducing
this feature.